Fix authentication issues and refactor routes

+++ todo-app/server/controllers/todoController.js
@@ -2,9 +2,11 @@ const Todo = require("../models/Todo");
 
 exports.createTodo = async (req, res) => {
   const { title } = req.body;
+  const userId = req.user.id;
 
   try {
     const newTodo = new Todo({
+      user: userId,
       title,
     });
 
@@ -17,8 +19,10 @@ exports.createTodo = async (req, res) => {
 };
 
 exports.getTodos = async (req, res) => {
+  const userId = req.user.id;
+
   try {
-    const todos = await Todo.find();
+    const todos = await Todo.find({ user: userId });
     res.json(todos);
   } catch (err) {
     console.error(err.message);
@@ -32,6 +36,10 @@ exports.updateTodo = async (req, res) => {
   try {
     let todo = await Todo.findById(req.params.id);
     if (!todo) return res.status(404).json({ msg: "Todo not found" });
+
+    if (todo.user.toString() !== req.user.id) {
+      return res.status(401).json({ msg: "Not authorized" });
+    }
 
     todo.title = title || todo.title;
     todo.completed = completed != null ? completed : todo.completed;
@@ -48,6 +56,10 @@ exports.deleteTodo = async (req, res) => {
   try {
     let todo = await Todo.findById(req.params.id);
     if (!todo) return res.status(404).json({ msg: "Todo not found" });
+
+    if (todo.user.toString() !== req.user.id) {
+      return res.status(401).json({ msg: "Not authorized" });
+    }
 
     await Todo.findByIdAndDelete(req.params.id);
     res.json({ msg: "Todo removed" });
